STORM-5775 - Modified pydstat.py to return the percentage of the CPU usage for all or specific process in a 1 second interval/screeshot
STORM-5775 - Modified pydstat.py to return the percentage of the CPU usage for all or specific process in a 1 second interval/screeshot

STORM-5775 - Modified pydstat.py to comply with the coding style guide. Modified setup.py to update the patch version number to 1.0.2 .

STORM-5775. The interval for calculating the CPU usage of a certain process in pydstat.py is now a variable and not a constant. Is 1 by default.

STORM-5775 - Modified pydstat.py and cmd.py of pydstat library. Now cmd.py can take three options for which proccess id the user want to log, what the screenshot time interval can be and the frequency of producing the logs.

STORM-5775 - Modified pydstat.py and cmd.py of pydstat library. Got rid of shlex, not neccessary for when arguments are escaped.

Bumping version of pydstat into 1.0.3.

PEP8 driven changes made to cmd.py and pydstat.py

Pydstat error handling change to allow cron to run the script with defaults without showing help menu. Bumped the pydstat version into 1.1.0.